More 32 bit to 64 bit changes

+++ libraries/app/include/graphene/app/api.hpp
@@ -181,9 +181,10 @@ namespace graphene { namespace app {
           * @return A list of operations performed by account, ordered from most recent to oldest.
           */
          vector<operation_history_object> get_relative_account_history( const std::string account_id_or_name,
-                                                                        uint32_t stop = 0,
+                                                                        uint64_t stop = 0,
                                                                         unsigned limit = 100,
-                                                                        uint32_t start = 0) const;
+                                                                        uint64_t start = 0) const;
+
          /**
           * @brief Get details of order executions occurred most recently in a trading pair
           * @param a One asset in a trading pair
